Add files via upload
NLP Task to predict the emotional intensity of the user who was written a particular tweet. 
I have used BOW (Bag of Words) technique to infer the emotional intensity of the user by taking into consideration the frequency of a certain set of words. 
Next file will most probably come with a much better accuracy score as I will be using Embedding techniques such as word-2-vec to get a better understanding of the semantic meaning behind certain words.